requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
jwt
python-jose
apscheduler